refactor(profile and useuser): create a component for the form and call useSession directly in useuser hook
