+++ app/profile/profile.module.css
@@ -1,14 +1,8 @@
-.container {
+.profile_wrapper {
+  width: 100%;
   display: flex;
   flex-direction: column;
-  gap: 20px;
-  height: 100%;
-}
-
-.buttons_wrapper{
-  display: flex;
-  width: 100%;
-  gap: 10px;
+  gap: 2rem;
 }
 
 .button {
@@ -20,29 +14,9 @@
   font-weight: var(--font-black);
 }
 
-.edit_button {
-  background-color: transparent;
-  border: 2px solid var(--primary-light-color);
-}
-
-.validate_button {
-  background-color: var(--primary-accent-color);
-}
-
 .logout_button {
   color: var(--danger-color);
   background-color: transparent;
   border: 2px solid var(--danger-color);
   margin-top: auto;
-}
-
-.display_creds {
-  background-color: transparent;
-  margin-right: auto;
-  color: var(--primary-light-color);
-  padding: 0;
-}
-
-.visible_button {
-  background-color: transparent;
 }
