+++ modules/client/requests/useUser.js
@@ -1,7 +1,9 @@
 import { useState, useEffect } from "react";
 import { useRouter } from "next/navigation";
+import { useSession, signOut } from "next-auth/react";
 
-export default function useUser(userSession) {
+export default function useUser() {
+  const { data: userSession } = useSession();
   const [userId, setUserId] = useState("");
   const [user, setUser] = useState();
   const baseUrl = process.env.NEXT_PUBLIC_API_URL;
